beatloop functie op de maat laten gaan

+++ Assets/Scripts/SongReader/ChartData.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 [System.Serializable]
 public class ChartData
 {
-    public float beat;
+    public int beat;
     public float[] headPos;
     public float[] leftHandPos;
     public float[] rightHandPos;
